Intentando de hacer funcionar POST a ControladorPedidos

+++ src/main/java/com/tallerwebi/repositorios/RepositorioPedido.java
@@ -3,7 +3,7 @@ package com.tallerwebi.repositorios;
 import com.tallerwebi.dominio.Pedido;
 
 public interface RepositorioPedido {
-    Pedido buscarPedido(int id);
+    Pedido buscarPedido(Long id);
     void guardar(Pedido pedido);
     void modificar(Pedido pedido);
 }
