+++ src/main/java/com/tallerwebi/servicios/ServicioPedido.java
@@ -5,7 +5,7 @@ import com.tallerwebi.dominio.Vehiculo;
 import com.tallerwebi.dominio.Viaje;
 
 public interface ServicioPedido {
-    Long agregarPedido(Vehiculo vehiculo, int pedidoId) throws Exception;
-    Pedido getPedido(int id);
+    Long agregarPedido(Vehiculo vehiculo, Long pedidoId) throws Exception;
+    Pedido getPedido(Long id);
 
 }
